+++ ruma-identifiers-validation/src/user_id.rs
@@ -14,6 +14,11 @@ pub fn validate(s: &str) -> Result<(NonZeroU8, bool), Error> {
 ///
 /// Returns an `Err` for invalid user ID localparts, `Ok(false)` for historical user ID localparts
 /// and `Ok(true)` for fully conforming user ID localparts.
+///
+/// With the `compat` feature enabled, this will also return `Ok(false)` for invalid user ID
+/// localparts. User IDs that don't even meet the historical user ID restrictions exist in the wild
+/// due to Synapse allowing them over federation. This will likely be fixed in an upcoming room
+/// version; see <https://github.com/matrix-org/matrix-doc/pull/2828>.
 pub fn localpart_is_fully_comforming(localpart: &str) -> Result<bool, Error> {
     // See https://matrix.org/docs/spec/appendices#user-identifiers
     let is_fully_conforming = localpart
@@ -23,9 +28,13 @@ pub fn localpart_is_fully_comforming(localpart: &str) -> Result<bool, Error> {
     // If it's not fully conforming, check if it contains characters that are also disallowed
     // for historical user IDs. If there are, return an error.
     // See https://matrix.org/docs/spec/appendices#historical-user-ids
+    #[cfg(not(feature = "compat"))]
     if !is_fully_conforming && localpart.bytes().any(|b| b < 0x21 || b == b':' || b > 0x7E) {
         Err(Error::InvalidCharacters)
     } else {
         Ok(is_fully_conforming)
     }
+
+    #[cfg(feature = "compat")]
+    Ok(is_fully_conforming)
 }